+++ src/Pim/Bundle/BaseConnectorBundle/JobLauncher/SimpleJobLauncher.php
@@ -30,6 +30,13 @@ class SimpleJobLauncher extends BaseSimpleJobLauncher
     /** @var ObjectManager */
     protected $objectManager;
 
+    /**
+     * @param JobRepositoryInterface          $jobRepository
+     * @param MassEditJobConfigurationFactory $jobConfigFactory
+     * @param ObjectManager                   $objectManager
+     * @param string                          $rootDir
+     * @param string                          $environment
+     */
     public function __construct(
         JobRepositoryInterface $jobRepository,
         MassEditJobConfigurationFactory $jobConfigFactory,